requirements.txt
bentoml
boto3
focal_loss
keras==2.10.0
numpy==1.21.6
opencv_python==4.7.0.72
pandas==1.3.5
Pillow
python-dotenv
scikit_image==0.19.3
scikit_learn==1.0.2
silence_tensorflow
tensorflow==2.10.0
tensorflow_addons
gdown
numba
